+++ plone/app/event/at/upgrades/event.py
@@ -1,8 +1,10 @@
 from Products.ATContentTypes.interfaces.event import IATEvent
+from Products.CMFCore.interfaces import IPropertiesTool
 from Products.CMFCore.utils import getToolByName
 from Products.contentmigration.archetypes import ATItemMigrator
 from Products.contentmigration.walker import CustomQueryWalker
 from transaction import savepoint
+from zope.component import queryUtility
 from zope.event import notify
 from zope.lifecycleevent import ObjectModifiedEvent
 
@@ -40,6 +42,16 @@ def callBefore(oldobj):
 
 
 def upgrade_step_1(context):
+    # switch linkintegrity temp off
+    ptool = queryUtility(IPropertiesTool)
+    site_props = getattr(ptool, 'site_properties', None)
+    linkintegrity = site_props.getProperty(
+        'enable_link_integrity_checks',
+        False
+    )
+    site_props.manage_changeProperties(enable_link_integrity_checks=False)
+
+    # do migration
     portal = getToolByName(context, 'portal_url').getPortalObject()
     walker = CustomQueryWalker(
         portal, PAEATMigrator,
@@ -47,4 +59,10 @@ def upgrade_step_1(context):
         callBefore=callBefore)
     savepoint(optimistic=True)
     walker.go()
+
+    # switch linkintegrity back
+    site_props.manage_changeProperties(
+        enable_link_integrity_checks=linkintegrity
+    )
+
     return walker.getOutput()